Update codebase to the most recent version of camps-dct

+++ faceplusplus/crawler.py
@@ -1,12 +1,8 @@
 # -*- coding: iso-8859-1 -*-
 
 import os
-import socket
 import json
-import time
 import common
-import random
-import fpp.apikey
 from fpp.facepp import API
 from fpp.facepp import APIError
 
@@ -17,54 +13,55 @@ class Crawler:
     
     def crawl(self, resourceID, filters):
         echo = common.EchoHandler(self.config)
-        echo.out(u"User ID received: %s." % resourceID)
+        
+        # Configura diretórios base para armazenamento
+        fppBaseDir = "../../data/fppselfies"
+        fppSubDir = str(int(resourceID.split("_")[0]) % 1000)
+        fppDataDir = os.path.join(fppBaseDir, fppSubDir)
+        if not os.path.exists(fppDataDir): os.makedirs(fppDataDir)
+        
+        # Inicializa variáveis de retorno
+        extraInfo = {"mediaerrors": [], "output": []}
+        
+        # Verifica se o arquivo já existe
+        fppFilePath = os.path.join(fppDataDir, "%s.json" % resourceID)
+        if os.path.isfile(fppFilePath): 
+            echo.out(u"Media %s alerady exists." % resourceID)
+            return (None, extraInfo, None)
         
         # Extrai filtros
-        application = filters[0]["data"]["application"]
+        imageURL = filters[0]["data"]["url"]
+        application = filters[1]["data"]["application"]
     
         # Constrói objeto da API com as credenciais de acesso
         apiServer = application["apiserver"]
         apiKey = application["apikey"]
         apiSecret = application["apisecret"]
-        api = API(srv = apiServer, key = apiKey, secret = apiSecret, timeout = 5, max_retries = 0, retry_delay = 0)
-        echo.out(u"App: %s." % str(application["name"]))
-    
-        # Configura diretórios base para armazenamento
-        fppBaseDir = "../../data/fpp"
-        fppDataDir = os.path.join(fppBaseDir, str(resourceID % 1000), str(resourceID))
-        if not os.path.exists(fppDataDir): os.makedirs(fppDataDir)
-        
-        # Carrega arquivo de feed do usuário
-        feedsBaseDir = "../../data/feeds"
-        feedsFilePath = os.path.join(feedsBaseDir, str(resourceID % 1000), "%s.feed" % resourceID)
-        with open(feedsFilePath, "r") as feedFile: feed = json.load(feedFile)
-        
-        # Determina mídias a serem coletadas
-        feedSampleSize = 10
-        feedList = random.sample(feed, min(len(feed), feedSampleSize))
-        
-        # Inicializa variáveis de retorno
-        extraInfo = {"mediaerrors": [], "usersok": [], "usersproblem": []}
+        api = API(srv = apiServer, key = apiKey, secret = apiSecret, timeout = 60, max_retries = 0, retry_delay = 0)
+        echo.out(u"ID: %s (App: %s)." % (resourceID, application["name"]))
         
         # Executa coleta
         attributes = ["gender", "age", "race", "smiling", "glass", "pose"]
-        for i, media in enumerate(feedList):
-            echo.out(u"Collecting media %d." % (i + 1))
-            while (True):
-                try:
-                    response = api.detection.detect(url = media["images"]["low_resolution"]["url"], attribute = attributes)
-                except Exception as error: 
-                    # Códigos de erro HTTP: http://www.faceplusplus.com/detection_detect/
-                    if isinstance(error, APIError): message = "%d: %s" % (error.code, json.loads(error.body)["error"])
-                    # socket.error e urllib2.URLError 
-                    else: message = str(error)
-                    extraInfo["mediaerrors"].append((media["id"], {"error": message}))
-                    extraInfo["usersproblem"].append((resourceID, None))
-                    return (None, extraInfo, None)
-                else:
-                    fppFilePath = os.path.join(fppDataDir, "%s.fpp" % media["id"])
-                    with open(fppFilePath, "w") as fppFile: json.dump(response, fppFile)
-                    break
-                    
-        extraInfo["usersok"].append((resourceID, None))
+        try:
+            response = api.detection.detect(url = imageURL, attribute = attributes)
+        except Exception as error: 
+            # Códigos de erro HTTP: http://www.faceplusplus.com/detection_detect/
+            if isinstance(error, APIError): message = "%d: %s" % (error.code, json.loads(error.body)["error"])
+            # socket.error e urllib2.URLError 
+            else: message = str(error)
+            extraInfo["mediaerrors"].append((resourceID, {"error": message}))
+        else: 
+            with open(fppFilePath, "w") as fppFile: json.dump(response, fppFile)
+            
+            for face in response["face"]:
+                faceInfo = filters[0]["data"]
+                faceInfo["gender_val"] = face["attribute"]["gender"]["value"]
+                faceInfo["gender_cnf"] = face["attribute"]["gender"]["confidence"]
+                faceInfo["race_val"] = face["attribute"]["race"]["value"]
+                faceInfo["race_cnf"] = face["attribute"]["race"]["confidence"]
+                faceInfo["smile"] = face["attribute"]["smiling"]["value"]
+                faceInfo["age_val"] = face["attribute"]["age"]["value"]
+                faceInfo["age_rng"] = face["attribute"]["age"]["range"]
+                extraInfo["output"].append((resourceID, faceInfo))
+        
         return (None, extraInfo, None)
